feat: lift the canonical serializer out of the supervisor
The audit log has always needed a byte-stable serialization, because a hash
chain over pretty-printed JSON is a chain over whatever order the keys happened
to come out in. That serializer lived inside the supervisor, where only the
supervisor could reach it.

A portable workflow needs the same guarantee for a different reason: a
fingerprint has to be the same number on the machine that wrote the file and the
machine that reads it, which is exactly the property the supervisor already
depends on. Sharing one implementation is what keeps the two from drifting into
two nearly-identical functions that disagree about a corner nobody tests.

No behaviour change. The supervisor's own tests are the proof.
